flask-dashboard/requirements.txt
# Flask and extensions
Flask==3.0.0
python-dotenv==1.0.0

# Azure (for future deployment)
azure-storage-blob==12.19.0

# Data handling
pandas==2.1.4
numpy==1.26.2

# ML / Predictions
scikit-learn==1.5.2
joblib==1.3.2
sbrscrape==0.0.9
beautifulsoup4==4.12.2
lxml==5.1.0
requests>=2.31.0

# Production server
gunicorn==21.2.0
